# backend/openapi.yaml
openapi: 3.0.3
info:
  title: IMS Backend API
  version: 1.0.0
  description: OpenAPI specification for the IMS backend. Covers auth, products, stock, supply, dispatch and reports routes.
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /products:
    get:
      summary: List products
      tags: [Products]
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
  /products/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a product by id
      tags: [Products]
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
    put:
      summary: Update a product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete a product
      tags: [Products]
      responses:
        '204':
          description: Product deleted
  /stock:
    get:
      summary: Get inventory stock list
      tags: [Stock]
      responses:
        '200':
          description: Stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItem'
    post:
      summary: Adjust or add stock
      tags: [Stock]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockItem'
      responses:
        '200':
          description: Stock adjusted
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
    StockItem:
      type: object
      properties:
        product_id:
          type: string
        location:
          type: string
        quantity:
          type: integer
...